+++ parser/build_cmd_structure.c
