+++ parser/build_cmd_structure_utils.c
