+++ parser/clear_structures.c
