+++ parser/set_cmd_attributes.c
