--- script1.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-echo "Just a script for you to test execution!"
